// frontend/src/components/units/map/mapkorea/MapKorea.types.ts
import { RefObject } from 'react';
export interface BarData {
    name: string;
    rating: number;
    reviewCount: number;
    address: string;
}


export interface MapKoreaPageUIProps {
    svgRef: RefObject<SVGSVGElement>;
    name: string;
    onClickMoveToMapPage: () => void;
    onClickMoveToMainPage: () => void;
    top10Bars: BarData[];
    onClickMoveToThisBar: (address: string) => void;
}